requirements.txt
numpy==1.24.2
tensorflow-gpu==2.10.1
pyyaml==6.0